requirements.txt
absl-py==2.0.0
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.12.1
asttokens==2.4.1
attrs==23.1.0
Babel==2.14.0
-e git+https://github.com/bnpy/bnpy.git@ffc2242427451aa6a61dcac1473c47577a5ade6f#egg=bnpy
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.2
cchardet==2.1.7
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==3.0.0
colorlog==6.7.0
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography @ file:///croot/cryptography_1694444244250/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.36
DataProperty==1.0.1
decorator==5.1.1
docutils==0.20.1
etils==1.5.2
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.1
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.10.0
glfw==2.6.3
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.1
grpcio==1.60.0
gtimer==1.0.0b5
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
huggingface-hub==0.19.4
huggingface-sb3==3.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.33.0
imageio-ffmpeg==0.4.9
imagesize==1.4.1
importlib-resources==6.1.1
ipython==8.22.1
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==0.14.1
jsonlines==4.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lxml==5.2.2
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1698692105134/work
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
-e git+https://github.com/juandelos/Meta-RL.git@2060806cd9b3380a2e4107384975b77b3246e9ea#egg=melts
memory-profiler==0.61.0
-e git+https://github.com/juandelos/multi-task-rl@da1c52983a70719a82bcf080506603e35e8c944c#egg=meta_envs&subdirectory=Meta_RL/submodules/meta-environments-main
-e git+https://github.com/juandelos/multi-task-rl@da1c52983a70719a82bcf080506603e35e8c944c#egg=meta_rand_envs&subdirectory=meta_rand_envs
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
-e git+https://github.com/juandelos/multi-task-rl@da1c52983a70719a82bcf080506603e35e8c944c#egg=mrl_analysis&subdirectory=Meta_RL/submodules/MRL-analysis-tools-main
msgpack==1.0.7
mujoco-py==2.1.2.14
munkres==1.0.12
networkx @ file:///croot/networkx_1690561992265/work
nose==1.3.7
numexpr @ file:///croot/numexpr_1696515281613/work
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
optuna==3.4.0
packaging @ file:///croot/packaging_1693575174725/work
pandas @ file:///croot/pandas_1697477158611/work/dist/pandas-2.1.1-cp310-cp310-linux_x86_64.whl#sha256=9030ae12668e518e702e668fa89f07f2847a175cdd526f95e39cb603d11ed9b8
parso==0.8.3
pathvalidate==3.2.0
pexpect==4.9.0
Pillow==10.1.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygame==2.5.2
pyglet==2.0.10
Pygments==2.17.2
PyOpenGL==3.1.7
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytablewriter==1.2.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1695131579487/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
-e git+https://github.com/juandelos/multi-task-rl@da1c52983a70719a82bcf080506603e35e8c944c#egg=rand_param_envs&subdirectory=rand_param_envs
ray==2.8.0
referencing==0.30.2
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rich==13.7.0
rl-zoo3==2.2.1
-e git+https://github.com/rail-berkeley/rlkit.git@c81509d982b4d52a6239e7bfe7d2540e3d3cd986#egg=rlkit
rpds-py==0.12.0
rsa==4.9
-e git+https://github.com/juandelos/multi-task-rl@da1c52983a70719a82bcf080506603e35e8c944c#egg=SAC&subdirectory=SAC
sb3-contrib==2.2.1
scikit-learn==1.0.2
scipy @ file:///croot/scipy_1696543286448/work/dist/scipy-1.11.3-cp310-cp310-linux_x86_64.whl#sha256=16a8c87e543daeb96d1310b2283f542fef2de84ee7090f60187eb71f71cb430c
seaborn==0.13.2
Shimmy==1.3.0
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
-e git+https://github.com/juandelos/multi-task-rl@da1c52983a70719a82bcf080506603e35e8c944c#egg=smrl&subdirectory=Meta_RL
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-gallery==0.15.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.23
stable-baselines3==2.2.1
stack-data==0.6.3
-e git+https://github.com/ElisevanderPol/symmetrizer.git@8dae02bee2ba7132ae4fb07e07020767d280842c#egg=symmetrizer
sympy @ file:///croot/sympy_1668202399572/work
tabledata==1.3.3
tcolorpy==0.1.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.14.1
triton==2.1.0
typepy==1.3.2
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1698257533958/work
wasabi==1.1.2
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
